+++ workdir/config/00_defines.lua
@@ -13,37 +13,9 @@ to_compress = {html=true, css=true, js=true, txt=true, php=true}
 content_type = {html="text/html", jpg="image/jpeg", png="image/png", gif="image/gif", txt="text/plain", css="text/css", ico="image/vnd.microsoft.icon", js="text/javascript",
 svg="image/svg+xml"}
 
--- ssl
--- default ssl cert is served when it doesn't match a vhost
-cert = create_cert ("workdir/ssl/cert.pem", "workdir/ssl/key.pem")
-
-if (cert) then
-  -- if you find a cert just use it
-  -- the server quits if you try to use the return of a failed create_cert
-  ssl_enable = true
-end
-
-
-main = add_vhost {
-host = {"localhost"},
-socket = {
-  {bind=nil, port="8080", sock_type="ipv4"},
-  {bind=nil, port="8081", sock_type="ipv4", ssl=true},
-},
-cert = cert,
-htdocs="htdocs",
---onRequest = default_request_handler,
---onError = default_error_handler,
---onFinish = default_finish_handler,
---hsts=600,
---hsts_flags="subdomains preload no_redirect no_header",
-}
-
 fpm_apps = {
 php = "tcp://localhost:9000",
 }
-
---do_proxy = create_round_robin ("http://localhost:8080/", "http://localhost:8080/", "http://localhost:8080/")
 
 --set_server_option (main, "directory_listing", true)
 --set_server_option (main, "disable", "compress")